% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{change_theme_one_button_page}
\alias{change_theme_one_button_page}
\title{Change the theme with a new page}
\usage{
change_theme_one_button_page(theme, body, content_background = NULL)
}
\arguments{
\item{content_background}{}
}
\description{
Change the theme with a new page
}
