% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certificate.R
\name{create_certificate}
\alias{create_certificate}
\title{Create certificate}
\usage{
create_certificate(name, overall, scores)
}
\arguments{
\item{scores}{}
}
\value{

}
\description{
Create certificate
}
