% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{list2grid}
\alias{list2grid}
\title{Arrange a list into a grid of fluidRows and columns}
\usage{
list2grid(li, ncol)
}
\arguments{
\item{li}{A list}

\item{ncol}{Number of columns}
}
\description{
Arrange a list into a grid of fluidRows and columns
}
