% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_test_feedback.R
\name{produce_3d_scatter}
\alias{produce_3d_scatter}
\title{Produce a 3D scatter graph of three scores}
\usage{
produce_3d_scatter(leaderboard, mdt_ability, bat_ability, mpt_ability)
}
\arguments{
\item{mpt_ability}{}
}
\description{
Produce a 3D scatter graph of three scores
}
