% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the ABCD study}
\usage{
run_app(
  title = "Discover Your Musical Genius!",
  admin_password = "demo",
  researcher_email = "D.Mullensiefen@gold.ac.uk",
  musicassessr_aws = FALSE,
  app_name = "ABCD",
  num_items = 5L
)
}
\arguments{
\item{num_items}{}
}
\description{
Run the ABCD study
}
