% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scp_down}
\alias{scp_down}
\title{Scp down}
\usage{
scp_down(
  file_to_down,
  download_loc = "/Users/sebsilas/Downloads",
  key = Sys.getenv("AWS_KEY_LOC")
)
}
\arguments{
\item{key}{}
}
\description{
Scp down
}
