% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{update_content_background}
\alias{update_content_background}
\title{Update the content background}
\usage{
update_content_background(background_hex)
}
\arguments{
\item{background_hex}{}
}
\description{
Update the content background
}
