% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{update_header_background}
\alias{update_header_background}
\title{Update the header}
\usage{
update_header_background(background_hex)
}
\arguments{
\item{background_hex}{}
}
\description{
Update the header
}
