% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{update_header_background_image}
\alias{update_header_background_image}
\title{Update an element}
\usage{
update_header_background_image(background_image)
}
\arguments{
\item{background_image}{}
}
\description{
Update an element
}
