% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certificate.R
\name{create_certificate}
\alias{create_certificate}
\title{Create certificate}
\usage{
create_certificate(
  name,
  overall_score,
  bat,
  bdt = NA,
  edt = NA,
  hpt = NA,
  mdt,
  mpt,
  msa = NA,
  mus_mellow = NA,
  mus_unpretentious = NA,
  mus_sophisticiated = NA,
  mus_intense = NA,
  piat = NA,
  rat = NA,
  saa = NA,
  tpt = NA
)
}
\arguments{
\item{overall_score}{}
}
\value{

}
\description{
Create certificate
}
