% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certificate.R
\name{create_certificate_filename}
\alias{create_certificate_filename}
\title{Create certificate filename}
\usage{
create_certificate_filename(name, overall_score)
}
\arguments{
\item{overall_score}{}
}
\value{

}
\description{
Create certificate filename
}
